requirements.txt
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
future==1.0.0
idna==3.7
Jinja2==3.1.4
junos-eznc==2.7.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
napalm==5.0.0
ncclient==0.6.15
netaddr==1.3.0
netmiko==4.3.0
netutils==1.8.1
nornir==3.4.1
nornir-jinja2==0.2.0
nornir-netbox==0.3.0
nornir-utils==0.2.0
nornir_napalm==0.5.0
ntc_templates==5.1.0
paramiko==3.4.0
pycparser==2.22
pyeapi==1.0.2
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scp==0.15.0
six==1.16.0
textfsm==1.1.3
transitions==0.9.1
ttp==0.9.5
ttp-templates==0.3.6
typing_extensions==4.12.2
urllib3==2.2.2
yamlordereddictloader==0.4.2
